pyproject.toml
[project]
name = "posts-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.0.0",
    "requests>=2.32.5",
    "python-dotenv>=1.0.0",
]
